Move header tests to their own file and expand them
